+++ Digipost.Api.Client.Domain/Message/MessageDeliveryResult.cs
